requirements.txt
Django==5.0.4
djangorestframework==3.15.1
gunicorn==22.0.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-decouple==3.8
django-filter==23.5
black==24.4.2